+++ src/upspring/ZipFile.cpp
@@ -21,7 +21,7 @@
 // --------------------------------------------------------------------------
 // Basic types.
 // --------------------------------------------------------------------------
-typedef unsigned long dword;
+typedef unsigned int dword;
 typedef unsigned short word;
 typedef unsigned char byte;
